okay lots of stuff - some isn't going to work as intended but placeholder
bookcase added along with bits as main homepage, idea of adding removing shelves (collections of books) via swipe to dismiss in principle this is the home page.

some refactors to reflect these changes.

need to review actions and state with viewmodel for this as a next phase get get them working correct (probably need to do the backend wiring to get data down).